+++ packages/onchainkit/src/connected/index.tsx
@@ -1,0 +1,1 @@
+export { Connected, ConnectedProps } from './components/Connected';
