+++ packages/onchainkit/src/index.ts
@@ -1,13 +1,14 @@
-// 🌲☀️🌲
+export { OnchainKitProvider } from './OnchainKitProvider';
+export { useOnchainKit } from './useOnchainKit';
+export { version } from './version';
+export type { OnchainKitProviderReact } from './types';
+
 export { isBase } from './core/utils/isBase';
 export { isEthereum } from './core/utils/isEthereum';
 export {
   getOnchainKitConfig,
   setOnchainKitConfig,
 } from './core/OnchainKitConfig';
-export { OnchainKitProvider } from './OnchainKitProvider';
-export { useOnchainKit } from './useOnchainKit';
-export { version } from './version';
 export type {
   AppConfig,
   IsBaseOptions,
@@ -16,4 +17,4 @@ export type {
   OnchainKitContextType,
 } from './core/types';
 
-export type { OnchainKitProviderReact } from './types';
+export { Connected } from './connected';
